Move NamedEntity and ReferenceID SegmentBridges to f27x
