+++ tigershark/facade/f27x.py
@@ -6,7 +6,9 @@ from tigershark.facade import D8
 from tigershark.facade import DR
 from tigershark.facade import TM
 from tigershark.facade import boolean
-from tigershark.facade.enums import date_or_time_qualifier
+from tigershark.facade.enum.common import date_or_time_qualifier
+from tigershark.facade.enum.common import reference_id_qualifier
+from tigershark.facade.enum.common import id_code_qualifier
 
 
 class Header(X12LoopBridge):
@@ -61,6 +63,48 @@ class TraceNumber(X12SegmentBridge):
     trace_number = ElementAccess("TRN", 2)
     entity_id = ElementAccess("TRN", 3)
     entity_additional_id = ElementAccess("TRN", 4)
+
+
+class NamedEntity(X12SegmentBridge):
+    entity_identifier = ElementAccess("NM1", 1, x12type=enum({
+            "03": "Dependent",
+            "1P": "Provider",
+            "2B": "Third-Party Administrator",
+            "36": "Employer",
+            "80": "Hospital",
+            "FA": "Facility",
+            "GP": "Gateway Provider",
+            "IL": "Insured",
+            "P5": "Plan Sponsor",
+            "PR": "Payer",
+            "QC": "Patient"}))
+    entity_type = ElementAccess("NM1", 2, x12type=enum({
+            "1": "Person",
+            "2": "Non-Person Entity"}))
+
+    last_name = ElementAccess("NM1", 3)
+    org_name = ElementAccess("NM1", 3)
+    first_name = ElementAccess("NM1", 4)
+    middle_initial = ElementAccess("NM1", 5)
+    suffix = ElementAccess("NM1", 7)
+
+    id_code = ElementAccess("NM1", 9)
+    id_code_qual = ElementAccess("NM1", 8, x12type=enum(id_code_qualifier))
+
+    @property
+    def is_person(self):
+        return self.entity_type[0] == "1"
+
+    @property
+    def is_organization(self):
+        return self.entity_type[0] == "2"
+
+
+class ReferenceID(X12SegmentBridge):
+    reference_id_qualifier = ElementAccess("REF", 1,
+            x12type=enum(reference_id_qualifier))
+    reference_id = ElementAccess("REF", 2)
+    description = ElementAccess("REF", 3)
 
 
 class ContactInformation(X12SegmentBridge):
